school/urls.py
from django.urls import path , re_path
from school import views
from rest_framework.urlpatterns import format_suffix_patterns

app_name = 'school'

courseView = views.CourseViewset.as_view({
    'post': 'create'
})

# Enter URL path below
urlpatterns = format_suffix_patterns([
    path('course/', courseView, name='course'),
])